requirements.txt
python-dotenv
pandas
openpyxl
python-dateutil
dicttoxml2
msoffcrypto-tool